final review fixes
