+++ lib/handlers/observeHandler.ts
@@ -5,7 +5,7 @@ import { LLMClient } from "../llm/LLMClient";
 import { StagehandPage } from "../StagehandPage";
 import { generateId } from "../utils";
 import { ScreenshotService } from "../vision";
-import { CDPSession } from "playwright";
+import { getAccessibilityTree, getXPathByResolvedObjectId } from "../a11y/utils";
 
 export class StagehandObserveHandler {
   private readonly stagehand: Stagehand;
@@ -236,219 +236,4 @@ export class StagehandObserveHandler {
     await this._recordObservation(instruction, elementsWithSelectors);
     return elementsWithSelectors;
   }
-}
-
-type AccessibilityNode = {
-  role: string;
-  name?: string;
-  description?: string;
-  value?: string;
-  children?: AccessibilityNode[];
-  childIds?: string[];
-  parentId?: string;
-  nodeId?: string;
-};
-
-interface TreeResult {
-  tree: AccessibilityNode[];
-  simplified: string;
-}
-
-// Parser function for str output
-function formatSimplifiedTree(node: AccessibilityNode, level = 0): string {
-  const indent = "  ".repeat(level);
-  let result = `${indent}[${node.nodeId}] ${node.role}${node.name ? `: ${node.name}` : ""}\n`;
-
-  if (node.children?.length) {
-    result += node.children
-      .map((child) => formatSimplifiedTree(child, level + 1))
-      .join("");
-  }
-  return result;
-}
-
-// Constructs the hierarchichal representation of the a11y tree
-function buildHierarchicalTree(nodes: AccessibilityNode[]): TreeResult {
-  const nodeMap = new Map<string, AccessibilityNode>();
-
-  // First pass: Create all important nodes
-  nodes.forEach((node) => {
-    const hasChildren = node.childIds && node.childIds.length > 0;
-    const hasValidName = node.name && node.name.trim() !== "";
-
-    // Skip nodes that have no name and no children
-    if (!hasValidName && !hasChildren) {
-      return;
-    }
-
-    nodeMap.set(node.nodeId, {
-      role: node.role,
-      nodeId: node.nodeId,
-      ...(hasValidName && { name: node.name }),
-      ...(node.description && { description: node.description }),
-      ...(node.value && { value: node.value }),
-    });
-  });
-
-  // Second pass: Build parent-child relationships
-  nodes.forEach((node) => {
-    if (node.parentId && nodeMap.has(node.nodeId)) {
-      const parentNode = nodeMap.get(node.parentId);
-      const currentNode = nodeMap.get(node.nodeId);
-
-      if (parentNode && currentNode) {
-        if (!parentNode.children) {
-          parentNode.children = [];
-        }
-        parentNode.children.push(currentNode);
-      }
-    }
-  });
-  // console.log(nodeMap);
-
-  nodes
-    .filter((node) => !node.parentId && nodeMap.has(node.nodeId))
-    .map((node) => nodeMap.get(node.nodeId))
-    .filter(Boolean) as AccessibilityNode[];
-
-  // Third pass: Clean up generic and none nodes by lifting their children
-  function cleanStructuralNodes(
-    node: AccessibilityNode,
-  ): AccessibilityNode | null {
-    if (!node.children) {
-      return node.role === "generic" || node.role === "none" ? null : node;
-    }
-
-    const cleanedChildren = node.children
-      .map((child) => cleanStructuralNodes(child))
-      .filter(Boolean) as AccessibilityNode[];
-
-    if (node.role === "generic" || node.role === "none") {
-      return cleanedChildren.length === 1
-        ? cleanedChildren[0]
-        : cleanedChildren.length > 1
-          ? { ...node, children: cleanedChildren }
-          : null;
-    }
-
-    return cleanedChildren.length > 0
-      ? { ...node, children: cleanedChildren }
-      : node;
-  }
-
-  const finalTree = nodes
-    .filter((node) => !node.parentId && nodeMap.has(node.nodeId))
-    .map((node) => nodeMap.get(node.nodeId))
-    .filter(Boolean)
-    .map((node) => cleanStructuralNodes(node))
-    .filter(Boolean) as AccessibilityNode[];
-
-  const simplifiedFormat = finalTree
-    .map((node) => formatSimplifiedTree(node))
-    .join("\n");
-
-  return {
-    tree: finalTree,
-    simplified: simplifiedFormat,
-  };
-}
-
-interface AXNode {
-  role?: { value: string };
-  name?: { value: string };
-  description?: { value: string };
-  value?: { value: string };
-  nodeId: string;
-  parentId?: string;
-  childIds?: string[];
-}
-
-async function getAccessibilityTree(
-  page: StagehandPage,
-  logger: (logLine: LogLine) => void,
-) {
-  await page.enableCDP("Accessibility");
-
-  try {
-    const { nodes } = await page.sendCDP<{ nodes: AXNode[] }>(
-      "Accessibility.getFullAXTree",
-    );
-
-    // Extract specific sources
-    const sources = nodes.map((node) => ({
-      role: node.role?.value,
-      name: node.name?.value,
-      description: node.description?.value,
-      value: node.value?.value,
-      nodeId: node.nodeId,
-      parentId: node.parentId,
-      childIds: node.childIds,
-    }));
-    // Transform into hierarchical structure
-    const hierarchicalTree = buildHierarchicalTree(sources);
-
-    return hierarchicalTree;
-  } catch (error) {
-    logger({
-      category: "observation",
-      message: "Error getting accessibility tree",
-      level: 1,
-      auxiliary: {
-        error: {
-          value: error.message,
-          type: "string",
-        },
-        trace: {
-          value: error.stack,
-          type: "string",
-        },
-      },
-    });
-    throw error;
-  } finally {
-    await page.disableCDP("Accessibility");
-  }
-}
-
-async function getXPathByResolvedObjectId(
-  cdpClient: CDPSession,
-  resolvedObjectId: string,
-): Promise<string> {
-  const { result } = await cdpClient.send("Runtime.callFunctionOn", {
-    objectId: resolvedObjectId,
-    functionDeclaration: `function() {
-      function getNodePath(node) {
-        const parts = [];
-        let current = node;
-        
-        while (current && current.parentNode) {
-          if (current.nodeType === Node.ELEMENT_NODE) {
-            let tagName = current.tagName.toLowerCase();
-            let sameTagSiblings = Array.from(current.parentNode.children).filter(
-              child => child.tagName === current.tagName
-            );
-            
-            if (sameTagSiblings.length > 1) {
-              let index = 1;
-              for (let sibling of sameTagSiblings) {
-                if (sibling === current) break;
-                index++;
-              }
-              tagName += '[' + index + ']';
-            }
-            
-            parts.unshift(tagName);
-          }
-          current = current.parentNode;
-        }
-        
-        return '/' + parts.join('/');
-      }
-      
-      return getNodePath(this);
-    }`,
-    returnByValue: true,
-  });
-
-  return result.value || "";
 }
