+++ types/context.ts
@@ -1,0 +1,25 @@
+export interface AXNode {
+  role?: { value: string };
+  name?: { value: string };
+  description?: { value: string };
+  value?: { value: string };
+  nodeId: string;
+  parentId?: string;
+  childIds?: string[];
+}
+
+export type AccessibilityNode = {
+  role: string;
+  name?: string;
+  description?: string;
+  value?: string;
+  children?: AccessibilityNode[];
+  childIds?: string[];
+  parentId?: string;
+  nodeId?: string;
+};
+
+export interface TreeResult {
+  tree: AccessibilityNode[];
+  simplified: string;
+}
